Log Crouch
Added a 3rd virtual camera to view the player moving through the log called "Upcam". This camera modifies an animator and created a new camera script.
Created a trigger game object for the log.
playermovement was modified to slow the player and elements of the playermovement where move to the camswitch script to make this possible. New game objects where created to represent the log.
Tags were added to game objects.